autogenerates password in the postbody
